link and compilation small fixes

+++ tools/WindowsUtils/FileUtils.cpp
@@ -2,7 +2,7 @@
 
 string getDirectory(const string& pFilename)
 {
-	int last = pFilename.find_last_of('/', '\\');
+	size_t last = pFilename.find_last_of('/', '\\');
 	if (last < 0)
 		return string("");
 	return string(pFilename.substr(0, last + 1));
